refactor: split layout types
