+++ packages/types/src/page/layouts/index.d.ts
@@ -1,0 +1,7 @@
+export { bindPageComponentSchema } from "./shared";
+export * from "./header";
+export * from "./footer";
+export * from "./section";
+export * from "./multi-column";
+export * from "./tabs";
+//# sourceMappingURL=index.d.ts.map
