+++ packages/types/src/page/layouts/index.ts
@@ -1,0 +1,6 @@
+export { bindPageComponentSchema } from "./shared";
+export * from "./header";
+export * from "./footer";
+export * from "./section";
+export * from "./multi-column";
+export * from "./tabs";
