+++ packages/types/src/page/layouts/multi-column.d.ts
@@ -1,0 +1,122 @@
+import { z } from "zod";
+import { type PageComponentBase } from "../base";
+import type { PageComponent } from "../page";
+export interface MultiColumnComponent extends PageComponentBase {
+    type: "MultiColumn";
+    columns?: number;
+    gap?: string;
+    children?: PageComponent[];
+}
+export declare const multiColumnComponentSchema: z.ZodObject<{
+    id: z.ZodString;
+    minItems: z.ZodOptional<z.ZodNumber>;
+    maxItems: z.ZodOptional<z.ZodNumber>;
+    desktopItems: z.ZodOptional<z.ZodNumber>;
+    tabletItems: z.ZodOptional<z.ZodNumber>;
+    mobileItems: z.ZodOptional<z.ZodNumber>;
+    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
+    href: z.ZodOptional<z.ZodString>;
+    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
+} & {
+    width: z.ZodOptional<z.ZodString>;
+    widthDesktop: z.ZodOptional<z.ZodString>;
+    widthTablet: z.ZodOptional<z.ZodString>;
+    widthMobile: z.ZodOptional<z.ZodString>;
+    height: z.ZodOptional<z.ZodString>;
+    heightDesktop: z.ZodOptional<z.ZodString>;
+    heightTablet: z.ZodOptional<z.ZodString>;
+    heightMobile: z.ZodOptional<z.ZodString>;
+} & {
+    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
+    top: z.ZodOptional<z.ZodString>;
+    left: z.ZodOptional<z.ZodString>;
+} & {
+    margin: z.ZodOptional<z.ZodString>;
+    marginDesktop: z.ZodOptional<z.ZodString>;
+    marginTablet: z.ZodOptional<z.ZodString>;
+    marginMobile: z.ZodOptional<z.ZodString>;
+    padding: z.ZodOptional<z.ZodString>;
+    paddingDesktop: z.ZodOptional<z.ZodString>;
+    paddingTablet: z.ZodOptional<z.ZodString>;
+    paddingMobile: z.ZodOptional<z.ZodString>;
+} & {
+    type: z.ZodLiteral<"MultiColumn">;
+    columns: z.ZodOptional<z.ZodNumber>;
+    gap: z.ZodOptional<z.ZodString>;
+    children: z.ZodDefault<z.ZodArray<z.ZodLazy<z.ZodTypeAny>, "many">>;
+}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
+    id: z.ZodString;
+    minItems: z.ZodOptional<z.ZodNumber>;
+    maxItems: z.ZodOptional<z.ZodNumber>;
+    desktopItems: z.ZodOptional<z.ZodNumber>;
+    tabletItems: z.ZodOptional<z.ZodNumber>;
+    mobileItems: z.ZodOptional<z.ZodNumber>;
+    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
+    href: z.ZodOptional<z.ZodString>;
+    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
+} & {
+    width: z.ZodOptional<z.ZodString>;
+    widthDesktop: z.ZodOptional<z.ZodString>;
+    widthTablet: z.ZodOptional<z.ZodString>;
+    widthMobile: z.ZodOptional<z.ZodString>;
+    height: z.ZodOptional<z.ZodString>;
+    heightDesktop: z.ZodOptional<z.ZodString>;
+    heightTablet: z.ZodOptional<z.ZodString>;
+    heightMobile: z.ZodOptional<z.ZodString>;
+} & {
+    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
+    top: z.ZodOptional<z.ZodString>;
+    left: z.ZodOptional<z.ZodString>;
+} & {
+    margin: z.ZodOptional<z.ZodString>;
+    marginDesktop: z.ZodOptional<z.ZodString>;
+    marginTablet: z.ZodOptional<z.ZodString>;
+    marginMobile: z.ZodOptional<z.ZodString>;
+    padding: z.ZodOptional<z.ZodString>;
+    paddingDesktop: z.ZodOptional<z.ZodString>;
+    paddingTablet: z.ZodOptional<z.ZodString>;
+    paddingMobile: z.ZodOptional<z.ZodString>;
+} & {
+    type: z.ZodLiteral<"MultiColumn">;
+    columns: z.ZodOptional<z.ZodNumber>;
+    gap: z.ZodOptional<z.ZodString>;
+    children: z.ZodDefault<z.ZodArray<z.ZodLazy<z.ZodTypeAny>, "many">>;
+}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
+    id: z.ZodString;
+    minItems: z.ZodOptional<z.ZodNumber>;
+    maxItems: z.ZodOptional<z.ZodNumber>;
+    desktopItems: z.ZodOptional<z.ZodNumber>;
+    tabletItems: z.ZodOptional<z.ZodNumber>;
+    mobileItems: z.ZodOptional<z.ZodNumber>;
+    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
+    href: z.ZodOptional<z.ZodString>;
+    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
+} & {
+    width: z.ZodOptional<z.ZodString>;
+    widthDesktop: z.ZodOptional<z.ZodString>;
+    widthTablet: z.ZodOptional<z.ZodString>;
+    widthMobile: z.ZodOptional<z.ZodString>;
+    height: z.ZodOptional<z.ZodString>;
+    heightDesktop: z.ZodOptional<z.ZodString>;
+    heightTablet: z.ZodOptional<z.ZodString>;
+    heightMobile: z.ZodOptional<z.ZodString>;
+} & {
+    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
+    top: z.ZodOptional<z.ZodString>;
+    left: z.ZodOptional<z.ZodString>;
+} & {
+    margin: z.ZodOptional<z.ZodString>;
+    marginDesktop: z.ZodOptional<z.ZodString>;
+    marginTablet: z.ZodOptional<z.ZodString>;
+    marginMobile: z.ZodOptional<z.ZodString>;
+    padding: z.ZodOptional<z.ZodString>;
+    paddingDesktop: z.ZodOptional<z.ZodString>;
+    paddingTablet: z.ZodOptional<z.ZodString>;
+    paddingMobile: z.ZodOptional<z.ZodString>;
+} & {
+    type: z.ZodLiteral<"MultiColumn">;
+    columns: z.ZodOptional<z.ZodNumber>;
+    gap: z.ZodOptional<z.ZodString>;
+    children: z.ZodDefault<z.ZodArray<z.ZodLazy<z.ZodTypeAny>, "many">>;
+}, z.ZodTypeAny, "passthrough">>;
+//# sourceMappingURL=multi-column.d.ts.map
