+++ packages/types/src/page/layouts/shared.d.ts
@@ -1,0 +1,4 @@
+import { z } from "zod";
+export declare let pageComponentSchemaRef: z.ZodTypeAny;
+export declare const bindPageComponentSchema: (schema: z.ZodTypeAny) => void;
+//# sourceMappingURL=shared.d.ts.map
