+++ packages/types/src/page/layouts/shared.ts
@@ -1,0 +1,6 @@
+import { z } from "zod";
+
+export let pageComponentSchemaRef: z.ZodTypeAny;
+export const bindPageComponentSchema = (schema: z.ZodTypeAny) => {
+  pageComponentSchemaRef = schema;
+};
